requirements.txt
kafka-python==2.0.2
python-json-logger==2.0.7